% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freedom-package.R
\docType{data}
\name{category_scores}
\alias{category_scores}
\title{Freedom in the World Dissaggregated Scores, 2012-2022}
\format{
A \code{data frame} with 62,235 observations and 8 variables.
}
\source{
Adapted from Freedom House.
}
\usage{
category_scores
}
\description{
Adding to the \code{country_rating_statuses} table, the items displayed here can be summed up
to obtain the \code{political_rights} aggregate score (sum of items A to C) and
\code{civil_liberties} aggregate score (items D to G). The sub-items are scored from 0 to 4,
with 0 representing the lowest degree of achievement in a given category and 4 the highest.
For example, a country that sums 36 or more points in the political rights aggregate score
obtains a political rights rating of 1, from 30 to 35 points a rating of 2, and so on.
}
\section{Variables}{


\itemize{
\item \code{year}: Year of observation (2012-2022).
\item \code{country_territory}: Country or territory name.
\item \code{iso2c}: ISO 2-character country code. Observations with \code{NA} could not be
matched with the \code{country_rating_statuses} table.
\item \code{iso3c}: ISO 3-character country code. Observations with \code{NA} could not be
matched with the \code{country_rating_statuses} table.
\item \code{continent}: Continent name. Observations with \code{NA} could not be
matched with the \code{country_rating_statuses} table.
\item \code{item}: Item letter (A-G).
\item \code{sub_item}: Sub-item letter and number (A1-G4).
\item \code{score}: Score for the sub-item (0-4).
}
}

\keyword{datasets}
