% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freedom-package.R
\docType{data}
\name{country_exports_similarity}
\alias{country_exports_similarity}
\title{Trade Network Coloured by Freedom Status}
\format{
A \code{igraph} object with 190 vertices (nodes) and 316 edges (arcs).
}
\source{
Adapted from the United Nations (trade volumes) and Freedom House (freedom information).
}
\usage{
country_exports_similarity
}
\description{
Connections between countries correspond to the strongest arcs based on the products they export.
The network was trimmed until obtaining an average of four arcs per node. This network was
obtained by using UN COMTRADE data for exports, using reports from importing countries.
}
\keyword{datasets}
