% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freedomhouse-package.R
\docType{data}
\name{country_rating_text}
\alias{country_rating_text}
\title{Freedom in the World Sub-Items Texts, 2017-2023}
\format{
A \code{data frame} with 36,093 observations and 10 variables.
}
\source{
Own creation, based on texts scraped from Freedom House.
}
\usage{
country_rating_text
}
\description{
Provides the text for each sub-item in the \code{country_scores} table. For each
sub-item the text corresponds to the justification for the assigned score.
}
\section{Variables}{


\itemize{
\item \code{year}: Year of observation (2017-2023).
\item \code{country}: Country name.
\item \code{iso2c}: ISO 2-character country code. Abkhazia, Crimea, Eastern Donbas, Kosovo,
Micronesia, Nagorno-Karabakh, Somaliland, South Ossetia, Tibet, Transnistria do not have
unambiguous matches and appear as 'NA'.
\item \code{iso3c}: ISO 3-character country code. Abkhazia, Crimea, Eastern Donbas, Kosovo,
Micronesia, Nagorno-Karabakh, Somaliland, South Ossetia, Tibet, Transnistria do not have
unambiguous matches and appear as 'NA'.
\item \code{continent}: Continent name.
\item \code{item}: Item letter (A-G).
\item \code{sub_item}: Sub-item letter and number (A1-G4) besides 'Overview', 'Key Developments'
'Rating Change', 'Trend Arrow', and 'Notes'.
\item \code{item_description}: Description of the item.
\item \code{sub_item_description}: Description of the sub-item.
\item \code{detail}: Details and justification for the assigned score in the sub-item.
}
}

\references{
Freedom House. 2023. Freedom in the World 2023: Marking 50 Years in the Struggle for Democracy.
Washington, DC: Freedom House. https://freedomhouse.org/report/freedom-world/2023/marking-50-years.
Accessed 2024-09-08.
}
\keyword{datasets}
