% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freedom-package.R
\docType{data}
\name{country_ratings_texts}
\alias{country_ratings_texts}
\title{Freedom in the World Sub-Items Texts, 2012-2022}
\format{
A \code{data frame} with 27,803 observations and 7 variables.
}
\source{
Own creation, based on texts scraped from Freedom House.
}
\usage{
country_ratings_texts
}
\description{
Provides the text for each sub-item in the \code{category_scores} table. For each
sub-item the text corresponds to the justification for the assigned score.
}
\section{Variables}{


\itemize{
\item \code{year}: Year of observation (2017-2022).
\item \code{country}: Country name.
\item \code{iso2c}: ISO 2-character country code. Abkhazia, Crimea, Eastern Donbas, Kosovo,
Micronesia, Nagorno-Karabakh, Somaliland, South Ossetia, Tibet, Transnistria do not have
unambiguous matches and appear as 'NA'.
\item \code{iso3c}: ISO 3-character country code. Abkhazia, Crimea, Eastern Donbas, Kosovo,
Micronesia, Nagorno-Karabakh, Somaliland, South Ossetia, Tibet, Transnistria do not have
unambiguous matches and appear as 'NA'.
\item \code{continent}: Continent name.
\item \code{sub_item}: Sub-item letter and number (A1-G4).
\item \code{detail}: Details and justification for the assigned score in the sub-item.
}
}

\references{
Freedom House. 2023. Freedom in the World 2023: Marking 50 Years in the Struggle for Democracy.
Washington, DC: Freedom House. https://freedomhouse.org/report/freedom-world/2023/marking-50-years.
Accessed 2022-10-01.
}
\keyword{datasets}
