% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freedomhouse-package.R
\docType{data}
\name{country_score}
\alias{country_score}
\title{Freedom in the World Dissaggregated Scores, 2012-2023}
\format{
A \code{data frame} with 62,875 observations and 10 variables.
}
\source{
Adapted from Freedom House.
}
\usage{
country_score
}
\description{
Adding to the \code{country_rating_statuses} table, the items displayed here can be summed up
to obtain the \code{political_rights} aggregate score (sum of items A to C) and
\code{civil_liberties} aggregate score (items D to G). The sub-items are scored from 0 to 4,
with 0 representing the lowest degree of achievement in a given category and 4 the highest.
For example, a country with 36 or more points in the political rights aggregate score
obtains a political rights rating of 1, from 30 to 35 points a rating of 2, and so on.
}
\section{Variables}{


\itemize{
\item \code{year}: Year of observation (2012-2023).
\item \code{country_territory}: Country or territory name.
\item \code{iso2c}: ISO 2-character country code. Abkhazia, Crimea, Eastern Donbas, Kosovo,
Micronesia, Nagorno-Karabakh, Somaliland, South Ossetia, Tibet, Transnistria do not have
unambiguous matches and appear as 'NA'.
\item \code{iso3c}: ISO 3-character country code. Same criteria for 'NA' as in \code{iso2c}.
\item \code{continent}: Continent name.
\item \code{item}: Item letter (A-G).
\item \code{sub_item}: Sub-item letter and number (A1-G4).
\item \code{item_description}: Description of the item.
\item \code{sub_item_description}: Description of the sub-item.
\item \code{score}: Score for the sub-item (0-4).
}
}

\keyword{datasets}
