% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freedomhouse-package.R
\docType{data}
\name{category_scores_items}
\alias{category_scores_items}
\title{Freedom in the World Items and Sub-Items Description, 2012-2022}
\format{
A \code{data frame} with 25 observations and 4 variables.
}
\source{
Adapted from Freedom House.
}
\usage{
category_scores_items
}
\description{
Provides a full description for the meaning of each item and sub-item in the
\code{category_scores} table. For example, item A corresponds to Political Rights.
}
\section{Variables}{


\itemize{
\item \code{item}: Item letter (A-G).
\item \code{sub_item}: Sub-item letter and number (A1-G4).
\item \code{item_description}: Description of the item.
\item \code{sub_item_description}: Description of the sub-item.
}
}

\references{
Freedom House. 2023. Freedom in the World 2023: Marking 50 Years in the Struggle for Democracy.
Washington, DC: Freedom House. https://freedomhouse.org/report/freedom-world/2023/marking-50-years.
Accessed 2022-10-01.
}
\keyword{datasets}
